backend/requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2
asyncpg
python-jose[cryptography]
passlib[bcrypt]
pydantic
pydantic-settings
python-multipart
qrcode[pil]
